Pull default extractor into separate file
